fix: update flows. set sensor classes and units. Rename files
