+++ custom_components/hsem/flows/energidataservice.py
@@ -40,7 +40,7 @@ def validate_energidataservice_input(user_input):
     ]
 
     for field in required_fields:
-        if not user_input.get(field):
+        if field not in user_input:
             errors[field] = "required"
 
     return errors
