+++ custom_components/hsem/flows/ev.py
@@ -41,7 +41,7 @@ def validate_ev_step_input(user_input):
     ]
 
     for field in required_fields:
-        if not user_input.get(field):
+        if field not in user_input:
             errors[field] = "required"
 
     return errors
