+++ custom_components/hsem/flows/huawei_solar.py
@@ -142,7 +142,7 @@ def validate_huawei_solar_input(user_input):
     ]
 
     for field in required_fields:
-        if not user_input.get(field):
+        if field not in user_input:
             errors[field] = "required"
 
     return errors
