+++ custom_components/hsem/flows/init.py
@@ -35,7 +35,7 @@ def validate_init_step_input(user_input):
     ]
 
     for field in required_fields:
-        if not user_input.get(field):
+        if field not in user_input:
             errors[field] = "required"
 
     return errors
