+++ custom_components/hsem/flows/weighted_values.py
@@ -102,7 +102,7 @@ def validate_weighted_values_input(user_input):
     ]
 
     for field in required_fields:
-        if not user_input.get(field):
+        if field not in user_input:
             errors[field] = "required"
 
     total_weight = (
